+++ pyproject.toml
@@ -1,0 +1,31 @@
+[build-system]
+requires = [ "setuptools>=61.2",]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "strain_tools"
+version = "0.1.0"
+description = "Tools for deriving surface-parallel strain rates and directions from glacier velocity fields."
+keywords = [ "glaciology", "ice", "ice sheet", "glacier", "velocity", "dynamics","geospatial", "satellite"]
+classifiers = [ "Intended Audience :: Science/Research", "Programming Language :: Python :: 3", "License :: OSI Approved :: MIT License", "Operating System :: OS Independent", "Topic :: Scientific/Engineering", "Topic :: Scientific/Engineering :: GIS", "Topic :: Scientific/Engineering :: Image Processing",]
+requires-python = ">=3.10"
+dependencies = ["rioxarray", "numpy", "numba"]
+
+[[project.authors]]
+name = "Tom Chudley"
+email = "thomas.r.chudley@durham.ac.uk"
+
+[project.readme]
+file = "README.md"
+content-type = "text/markdown"
+
+[project.license]
+text = "MIT"
+
+[project.urls]
+Homepage = "https://github.com/trchudley/glacier-strain-tools"
+Repository = "https://github.com/trchudley/glacier-strain-tools"
+Issues = "https://github.com/trchudley/glacier-strain-tools/issues"
+
+[tool.setuptools.packages.find]
+where = ["src"]
