--- test_strain_tools.py
@@ -1,1 +1,0 @@
-import strain_tools
